Update and strengthen rubocop rules

+++ app/controllers/simulator_controller.rb
@@ -1,11 +1,13 @@
+# frozen_string_literal: true
+
 class SimulatorController < ApplicationController
   include SimulatorHelper
   include ActionView::Helpers::SanitizeHelper
 
-  before_action :authenticate_user!, only: [:create, :update, :edit,:update_image]
-  before_action :set_project, only: [:show, :embed, :embed, :update, :edit, :get_data,:update_image]
-  before_action :check_view_access, only: [:show,:embed,:get_data]
-  before_action :check_edit_access, only: [:edit,:update,:update_image]
+  before_action :authenticate_user!, only: %i[create update edit update_image]
+  before_action :set_project, only: %i[show embed embed update edit get_data update_image]
+  before_action :check_view_access, only: %i[show embed get_data]
+  before_action :check_edit_access, only: %i[edit update update_image]
   after_action :allow_iframe, only: :embed
 
   def self.policy_class
@@ -15,11 +17,11 @@ class SimulatorController < ApplicationController
   def show
     @logix_project_id = params[:id]
     @external_embed = false
-    render 'embed'
+    render "embed"
   end
 
   def edit
-    @project = Project.find_by(id:params[:id])
+    @project = Project.find_by(id: params[:id])
     @logix_project_id = params[:id]
     @projectName = @project.name
   end
@@ -30,7 +32,7 @@ class SimulatorController < ApplicationController
     @project = Project.find(params[:id])
     @author = @project.author_id
     @external_embed = true
-    render 'embed'
+    render "embed"
   end
 
   def get_data
@@ -52,9 +54,7 @@ class SimulatorController < ApplicationController
     @project.name = sanitize(params[:name])
     @project.save
 
-    if check_to_delete(params[:image])
-      File.delete(image_file)
-    end
+    File.delete(image_file) if check_to_delete(params[:image])
 
     render plain: "success"
   end
@@ -76,30 +76,28 @@ class SimulatorController < ApplicationController
     @project.image_preview = image_file
     @project.save
 
-    if check_to_delete(params[:image])
-      File.delete(image_file)
-    end
+    File.delete(image_file) if check_to_delete(params[:image])
 
     # render plain: simulator_path(@project)
     # render plain: user_project_url(current_user,@project)
-    redirect_to edit_user_project_url(current_user,@project)
+    redirect_to edit_user_project_url(current_user, @project)
   end
 
   private
-  def allow_iframe
-    response.headers.except! 'X-Frame-Options'
-  end
 
-  def set_project
-    @project = Project.find(params[:id])
-  end
+    def allow_iframe
+      response.headers.except! "X-Frame-Options"
+    end
 
-  def check_edit_access
-    authorize @project, :edit_access?
-  end
+    def set_project
+      @project = Project.find(params[:id])
+    end
 
-  def check_view_access
-    authorize @project, :view_access?
-  end
+    def check_edit_access
+      authorize @project, :edit_access?
+    end
 
+    def check_view_access
+      authorize @project, :view_access?
+    end
 end
